requirements.txt
pandas==0.24.2
numpy==1.14.5
six==1.12.0
pyhocon==0.3.51
scipy==1.1.0
tensorpack==0.9.4
torchvision==0.4.0
ray==0.7.2
tensorflow_gpu==1.12.0
opencv_python==4.1.0.25
Pillow==6.1.0
PyYAML==5.1.1
requests==2.22.0
setproctitle==1.1.10
psutil==5.6.3
matplotlib==2.2.4
